requirements.txt
parameterized==0.8.1
pytest==7.1.3
python-dotenv==0.21.0
requests==2.25.1
